Reduce complexity of edit status endpoint [#162850075]  On branch reduce-complexity  Changes to be committed: 	modified:   .coverage 	modified:   api/app.py 	deleted:    api/routes/admin.py 	modified:   api/routes/delete_incident.py 	modified:   api/routes/edit_incident.py 	modified:   api/routes/get_incidents.py
